requirements.txt
aiocron==1.6
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
croniter==1.0.15
discord-pretty-help==1.3.2
discord.py==1.7.3
future==0.18.2
idna==2.10
multidict==5.1.0
mysql-connector-python==8.0.25
natsort==7.1.1
pkg-resources==0.0.0
protobuf==3.17.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
riotwatcher==3.1.2
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
yarl==1.6.3
